Async problem resolved
